feat(sift): upgrade to v13.4.0 and refactor queries
BREAKING CHANGE: The latest version of sift makes it function more like MongoDB, specifically moving away from the
ability to define nested queries. MongoDB does this to create a boundary between the query language
and the data. In turn, this will break the ability to specify nested queries in MSON.

For example, previous to this change, you could specify the following query:
```
{
  fields: {
    firstName: {
      value: {
        $eq: "Jimmy"
      }
    }
  }
}
```
After this change, you'll need to refactor the query to be in the form of:
```
{
  "fields.firstName.value": {
    $eq: "Jimmy"
  }
}
```

Note: this affects any areas in MSON that support queries, including the `if` clause in an Action.

See https://github.com/crcn/sift.js/issues/206 & https://github.com/redgeoff/mson/issues/254 for more information

fix #254